bharatbot/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
azure-cognitiveservices-speech==1.38.0
azure-ai-translation-text==1.0.0b1
azure-search-documents==11.6.0b4
azure-ai-projects==1.0.0b7
azure-identity==1.17.1
openai==1.35.0
python-multipart==0.0.9
aiofiles==23.2.1
